+++ Assets/Scripts/Vanilla/GameContent/Bosses/TheGiant.cs
@@ -21,7 +21,6 @@ using PVZEngine.Level;
 using PVZEngine.Modifiers;
 using Tools;
 using UnityEngine;
-using static UnityEngine.EventSystems.EventTrigger;
 
 namespace MVZ2.GameContent.Bosses
 {
@@ -121,7 +120,7 @@ namespace MVZ2.GameContent.Bosses
                         {
                             other.PlaySound(VanillaSoundID.pacmanKill);
                         }
-                    } 
+                    }
                 }
                 return;
             }