+++ Assets/Scripts/Vanilla/GameContent/Detections/DispenserDetector.cs
@@ -4,7 +4,6 @@ using MVZ2.Vanilla.Entities;
 using MVZ2.Vanilla.Level;
 using PVZEngine.Entities;
 using UnityEngine;
-using static UnityEngine.EventSystems.EventTrigger;
 
 namespace MVZ2.GameContent.Detections
 {