+++ Assets/Scripts/Vanilla/GameContent/Detections/WitherDetector.cs
@@ -2,7 +2,6 @@
 using MVZ2.Vanilla.Level;
 using PVZEngine.Entities;
 using UnityEngine;
-using static UnityEngine.EventSystems.EventTrigger;
 
 namespace MVZ2.GameContent.Detections
 {