+++ Assets/Scripts/Vanilla/GameContent/Enemies/Chapter3/ReverseSatellite.cs
@@ -13,7 +13,6 @@ using PVZEngine.Entities;
 using PVZEngine.Level;
 using Tools;
 using UnityEngine;
-using static UnityEngine.EventSystems.EventTrigger;
 
 namespace MVZ2.GameContent.Enemies
 {